+++ source/Cosmos.Core/IOGroup/PCSpeaker.cs
@@ -15,17 +15,17 @@ namespace Cosmos.Core.IOGroup
         /// <summary>
         /// Gate IO port.
         /// </summary>
-        public readonly ushort Gate = 0x61;
+        public readonly int Gate = 0x61;
         // These two ports are shared with the PIT, so names are the same
         // IO Port 43
         /// <summary>
         /// Command register IO port.
         /// </summary>
-        public readonly ushort CommandRegister = 0x43;
+        public readonly int CommandRegister = 0x43;
         // IO Port 42
         /// <summary>
         /// Channel to data IO port.
         /// </summary>
-        public readonly ushort Channel2Data = 0x42;
+        public readonly int Channel2Data = 0x42;
     }
 }
