+++ source/Cosmos.Core/IOGroup/RTC.cs
@@ -12,10 +12,10 @@ namespace Cosmos.Core.IOGroup {
         /// <summary>
         /// Address IOPort.
         /// </summary>
-        public readonly ushort Address = 0x70;
+        public readonly int Address = 0x70;
         /// <summary>
         /// Data IOPort.
         /// </summary>
-        public readonly ushort Data = 0x71;
+        public readonly int Data = 0x71;
     }
 }
